Joining tests to the main file
